requirements.txt
azure-storage-blob==12.9.0
watchdog==3.0.0